Remove split_bounded; keep its discovery for the portfolio split
PR #270 should add exactly one new algorithm, the portfolio `split`. The
bounded `split_bounded` router was already dropped from the production
registry in the port; remove it entirely rather than keeping a deprecated
engine (recoverable from branch history).

The portfolio reuses split_bounded's bounded, amount-aware candidate
discovery. Move that cluster (find_candidate_paths, CandidateSearchConfig,
SplitMarketRead, default_anchor_tokens, the CandidatePathState/ScoredEdge
types, the CANDIDATE_* constants, and the discovery helpers) into a new
split_discovery module with its own doc; point split_exp at it. Delete
split_bounded.rs (engine, allocators, SplitBoundedAlgorithm, tests) and its
mod/export.

Test coverage equivalents: the discovery test that mattered (finds both
parallel pools, ranks by live simulation) moves to split_discovery as
discovery_finds_and_ranks_parallel_pools, plus an invalid-hop guard test;
the engine/allocator tests die with the engine. The portfolio's
never-loses-to-split_bounded test is replaced by
portfolio_output_near_two_pool_optimum (gross within 0.1% of the analytical
two-pool optimum via optimal_two_pool_output, on fee-free pools), keeping the
small-order no-loss and tight-timeout (now vs the single path) checks.

Docs: delete the split-bounded page and its nav entry; the split-comparison
page stays as the historical record with an adapted epilogue noting the
removal and where discovery now lives.
