+++ fynd-core/src/algorithm/mod.rs
@@ -22,10 +22,7 @@ pub mod bellman_ford;
 pub mod most_liquid;
 pub mod path_frank_wolfe;
 pub mod split;
-// Retained as an offline benchmark baseline and test reference; its production registry entry was
-// replaced by the portfolio split, so most of its surface is unused outside tests.
-#[allow(dead_code)]
-pub mod split_bounded;
+pub mod split_discovery;
 pub mod split_exp;
 #[allow(dead_code)]
 pub(crate) mod split_primitives;
@@ -41,7 +38,6 @@ pub use bellman_ford::BellmanFordAlgorithm;
 pub use most_liquid::MostLiquidAlgorithm;
 pub use path_frank_wolfe::PathFrankWolfeAlgorithm;
 pub use split::SplitAlgorithm;
-pub use split_bounded::SplitBoundedAlgorithm;
 pub use split_exp::ExpSplitAlgorithm;
 use tycho_simulation::tycho_core::models::Address;
